It confuse the unit mm / minute suddenly being mentioned in Default Max Acceleration
I propose changing the txt:
"A value of 3000 means that an axis may accelerate from 0 to 3000mm/m (50mm/s) within a one second movement."
in to txt:
"A value of e.g 50 means that an axis can accelerate from 0 to 50 mm/s in one second. 
It is to be noted that in G-code, speed unit is mm/m. (e.g 1500 mm/minute = 25 mm/s).
Intuitively, we can relate to acceleration as G-forces. 1G is roughly  10 m/^2 = 10000 mm/s^2.
An acleration value of 3000 [mm/^2] is 3 m/^2 (approximately 0.3G).