Add documentation for verbose level 3
There is no documentation for verbose level 3.  According to G33.cpp, range of 0-3 is accepted and 3 adds the printing of calibration results.  Also, there is `default: 1` defined for the V parameter, however that isn't shown on the actual documentation site (nothing clearly explicitly states that the default value is 1, other than the example below. Perhaps add some type of text to denote that the default verbose level is 1?