Tweak explanation of parameter E
The previous explanation of "E" made it sound like it was a relative parameter - that doing two consecutive moves with E5 would yield 10 units of extruded filament.  Remembering a different explanation, and checking some generated gcode, this appears to be incorrect, and the E parameter is effectively treated as an axis like the others.