Update G029-ubl.md
Changed the command to heat the bed in the optional settings from M190 to M140. The M190 waits for bed to finish heating until processing more commands where the M140 starts the heating process and continues on. This change makes the documentation consistent with the use of M104 to heat the hotend instead of M109 which waits until the hotend reaches the temperature specified.